Support webpack 2.1.0-beta as a peer dependency.
The range ">=1.12.9 <3.0.0" does not include any beta versions. Given webpack2 is currently in beta, it needs to be explicitly allowed to prevent a missing peer warning:
"npm WARN babel-plugin-webpack-loaders@0.8.0 requires a peer of webpack@>=1.12.9 <3.0.0 but none was installed."

While the warning doesn't stop npm install on npm@4, it does prevent shrinkwrap from running.